Extract `preprocess`
